+++ roles/monitoring_agent/files/scripts/ton-depool-ticktok.sh
@@ -7,7 +7,7 @@ TON_ELECTION_TICKTOK_IS_SENDED="depool.ticktock.sended"
 . ton-env.sh
 
 ton-check-env.sh TON_CLI
-ton-check-env.sh TON_DAPP
+ton-check-env.sh TON_CLI_CONFIG
 ton-check-env.sh DEPOOL_ADDR
 ton-check-env.sh TIK_ADDR
 ton-check-env.sh TIK_PRV_KEY
@@ -15,7 +15,7 @@ ton-check-env.sh TIK_PRV_KEY
 if [ $# == 1 ];then
    if [ $1 = '-f' ] || [ $1 = '-force' ];then
        echo "INFO: force mod"
-       $TON_CLI --url $TON_DAPP depool --addr $DEPOOL_ADDR ticktock -w $TIK_ADDR -s $TIK_PRV_KEY
+       $TON_CLI -c $TON_CLI_CONFIG depool --addr $DEPOOL_ADDR ticktock -w $TIK_ADDR -s $TIK_PRV_KEY
        exit
     fi
 fi
@@ -28,10 +28,17 @@ if [ $ELECTION_STATE != "ACTIVE" ];then
 fi
 
 # get elector address
-ELECTOR_ADDR="-1:$($TON_CLI --url $TON_DAPP  getconfig 1 | grep 'p1:' | sed 's/Config p1:[[:space:]]*//g' | tr -d \")"
+ELECTOR_ADDR="-1:$($TON_CLI -c $TON_CLI_CONFIG  getconfig 1 | grep 'p1:' | sed 's/Config p1:[[:space:]]*//g' | tr -d \")"
 
 # get elector start (unixtime)
-ELECTIONS_DATE=$($TON_CLI --url $TON_DAPP runget $ELECTOR_ADDR active_election_id  | grep 'Result:' | sed 's/Result:[[:space:]]*//g' | tr -d \"[])
+ELECTIONS_DATE=$($TON_CLI -c $TON_CLI_CONFIG runget $ELECTOR_ADDR active_election_id  | grep 'Result:' | sed 's/Result:[[:space:]]*//g' | tr -d \"[])
+
+if [ -z $ELECTIONS_DATE ]; then
+
+   ## hotfix try to use new solidity contract for rustnet.ton.dev
+   ELECTION_RESULT=`$TON_CLI -c $TON_CLI_CONFIG run $ELECTOR_ADDR active_election_id {} --abi $TON_CONTRACT_ELECTOR_ABI`
+   ELECTIONS_DATE=$(echo $ELECTION_RESULT | awk -F'Result: ' '{print $2}' | jq -r '.value0'  )
+fi
 
 TON_ELECTION_SUBFOLDER="$TON_ELECTION_FOLDER/$ELECTIONS_DATE"
 if [ ! -d $TON_ELECTION_SUBFOLDER ]; then
@@ -43,7 +50,7 @@ if [ -f $TON_ELECTION_SUBFOLDER/$TON_ELECTION_TICKTOK_IS_SENDED ]; then
    exit
 fi
    
-DEPOOL_TICKTOK_RESULT="$($TON_CLI --url $TON_DAPP depool --addr $DEPOOL_ADDR ticktock -w $TIK_ADDR -s $TIK_PRV_KEY)"
+DEPOOL_TICKTOK_RESULT="$($TON_CLI -c $TON_CLI_CONFIG depool --addr $DEPOOL_ADDR ticktock -w $TIK_ADDR -s $TIK_PRV_KEY)"
 
 DEPOOL_TICKTOK_TRANSACTION_ID="$(echo $DEPOOL_TICKTOK_RESULT | awk -F'Result: ' '{print $2}' | jq -r '.transId')"
 if [ -z "$DEPOOL_TICKTOK_TRANSACTION_ID" ];then